backend/pyproject.toml
[project]
name = "boilerworks-fastapi-nextjs-backend"
version = "0.1.0"
description = "Boilerworks FastAPI + Next.js template — async Python backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "redis>=5.0",
    "bcrypt>=4.2",
    "python-slugify>=8.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=5.0",
    "coverage>=7.6",
    "httpx>=0.28",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "N"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/tests/*", "alembic/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
